Move xml.py into RimSort.utils
