+++ RimSort/views/main_content_panel.py
@@ -72,7 +72,7 @@ from util.steam.steamworks.wrapper import (
     SteamworksSubscriptionHandler,
 )
 from RimSort.utils.todds.wrapper import ToddsInterface
-from util.xml import json_to_xml_write, xml_path_to_json
+from RimSort.utils.xml import json_to_xml_write, xml_path_to_json
 from RimSort.views.game_configuration_panel import GameConfiguration
 from RimSort.windows.missing_mods_panel import MissingModsPrompt
 from RimSort.windows.rule_editor_panel import RuleEditor
